# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Finance App API",
    "description": "API para gestão de fluxo de caixa, contas bancárias e investimentos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/transactions": {
      "get": {
        "tags": [
          "Transactions"
        ],
        "summary": "List Transactions",
        "operationId": "list_transactions_api_transactions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "type": "array",
                  "title": "Response List Transactions Api Transactions Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Transactions"
        ],
        "summary": "Create Transaction",
        "operationId": "create_transaction_api_transactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/transactions/{transaction_id}": {
      "patch": {
        "tags": [
          "Transactions"
        ],
        "summary": "Update Transaction",
        "description": "Edição parcial.\n\nNão há estorno nem reaplicação de saldo: com o saldo derivado do ledger, a\npróxima leitura simplesmente recalcula. O `_apply_to_balance` que existia\naqui — e a regra de ordenar as validações antes de encostar no saldo —\nperderam o objeto junto com a coluna.",
        "operationId": "update_transaction_api_transactions__transaction_id__patch",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Transactions"
        ],
        "summary": "Delete Transaction",
        "description": "Exclui a transação.\n\nSem estorno de saldo: a linha some do ledger e o saldo derivado acompanha.",
        "operationId": "delete_transaction_api_transactions__transaction_id__delete",
        "parameters": [
          {
            "name": "transaction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Transaction Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investments": {
      "get": {
        "tags": [
          "Investments"
        ],
        "summary": "List Investments",
        "operationId": "list_investments_api_investments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvestmentResponse"
                  },
                  "type": "array",
                  "title": "Response List Investments Api Investments Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Investments"
        ],
        "summary": "Create Investment",
        "operationId": "create_investment_api_investments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/investments/{investment_id}/history": {
      "post": {
        "tags": [
          "Investments"
        ],
        "summary": "Create Investment History",
        "operationId": "create_investment_history_api_investments__investment_id__history_post",
        "parameters": [
          {
            "name": "investment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestmentHistoryCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestmentHistoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Investments"
        ],
        "summary": "List Investment History",
        "operationId": "list_investment_history_api_investments__investment_id__history_get",
        "parameters": [
          {
            "name": "investment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Investment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InvestmentHistoryResponse"
                  },
                  "title": "Response List Investment History Api Investments  Investment Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryRef": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "color": {
            "type": "string",
            "title": "Color"
          },
          "icon_name": {
            "type": "string",
            "title": "Icon Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "color",
          "icon_name"
        ],
        "title": "CategoryRef",
        "description": "Categoria aninhada nas responses de transação e parcelamento.\n\nCarrega `color` e `icon_name` junto do nome para que o front pinte a badge\nsem uma segunda chamada a /categories. Serialização direta do ORM pela\nrelação — ver \"Design Patterns\" no CLAUDE.md; toda listagem que expõe este\ncampo precisa de `joinedload`, senão vira N+1."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InstallmentProgress": {
        "properties": {
          "current_installment": {
            "type": "integer",
            "title": "Current Installment"
          },
          "total_installments": {
            "type": "integer",
            "title": "Total Installments"
          }
        },
        "type": "object",
        "required": [
          "current_installment",
          "total_installments"
        ],
        "title": "InstallmentProgress"
      },
...
      "InvestmentCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "title": "Name",
            "description": "Nome do investimento"
          },
          "current_balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Current Balance",
            "description": "Aporte ou saldo inicial do investimento"
          }
        },
        "type": "object",
        "required": [
          "name",
          "current_balance"
        ],
        "title": "InvestmentCreate"
      },
      "InvestmentHistoryCreate": {
        "properties": {
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "description": "Data do registro de histórico"
          },
          "balance": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Balance",
            "description": "Saldo do investimento na data informada"
          }
        },
        "type": "object",
        "required": [
          "date",
          "balance"
        ],
        "title": "InvestmentHistoryCreate"
      },
      "InvestmentHistoryResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Balance"
          },
          "investment_id": {
            "type": "integer",
            "title": "Investment Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "date",
          "balance",
          "investment_id"
        ],
        "title": "InvestmentHistoryResponse"
      },
      "InvestmentResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "current_balance": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Current Balance"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "current_balance"
        ],
        "title": "InvestmentResponse"
      },
...
      "TransactionCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 150,
            "title": "Title",
            "description": "Título/descrição da transação"
          },
          "type": {
            "type": "string",
            "title": "Type",
            "description": "Tipo de transação: 'ENTRADA' ou 'SAÍDA'"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              }
            ],
            "title": "Amount",
            "description": "Valor da transação"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date",
            "description": "Data da transação"
          },
          "category_id": {
            "type": "integer",
            "title": "Category Id",
            "description": "ID da categoria cadastrada"
          },
          "is_fixed": {
            "type": "boolean",
            "title": "Is Fixed",
            "description": "Indica se é uma despesa fixa/recorrente",
            "default": false
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id",
            "description": "ID da conta bancária relacionada"
          },
          "installment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Id",
            "description": "ID do parcelamento de origem, se houver"
          }
        },
        "type": "object",
        "required": [
          "title",
          "type",
          "amount",
          "date",
          "category_id",
          "account_id"
        ],
        "title": "TransactionCreate"
      },
      "TransactionResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "category": {
            "$ref": "#/components/schemas/CategoryRef"
          },
          "is_fixed": {
            "type": "boolean",
            "title": "Is Fixed"
          },
          "account_id": {
            "type": "integer",
            "title": "Account Id"
          },
          "installment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Id"
          },
          "installment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/InstallmentProgress"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "type",
          "amount",
          "date",
          "category",
          "is_fixed",
          "account_id"
        ],
        "title": "TransactionResponse"
      },
      "TransactionUpdate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type",
            "description": "'ENTRADA' ou 'SAÍDA'"
          },
          "amount": {
            "anyOf": [
              {
                "type": "number",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "category_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "is_fixed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Fixed"
          },
          "installment_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Installment Id",
            "description": "Só aceita null (desvincular) — ver decisão B6"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "TransactionUpdate",
        "description": "Payload parcial do `PATCH /transactions/{id}`.\n\nTodo campo é opcional: o que não vier no corpo não é tocado. O router lê\n`model_dump(exclude_unset=True)` para distinguir \"não enviado\" de \"enviado\ncomo null\" — a diferença importa em `installment_id`, onde `null` é um\npedido explícito de desvínculo.\n\n`account_id` **não existe aqui de propósito** (decisão A3). Com\n`extra=\"forbid\"` a tentativa vira 422; sem ele, o Pydantic descartaria o\ncampo em silêncio e a resposta 200 faria o cliente acreditar que a\ntransação mudou de conta.\n\nA regra `is_fixed` × `installment_id` **não** está neste schema — ver\n`update_transaction` no router. Num payload parcial ela depende do estado\nmesclado com a linha do banco, que o schema não enxerga."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}